separated controllers and html
