+++ src/main/java/org/elevenfiftyconsulting/beans/Note.java
@@ -59,5 +59,31 @@ public class Note {
 	public void setModifiedUtc(OffsetDateTime modifiedUtc) {
 		ModifiedUtc = modifiedUtc;
 	}
+	@Override
+	public int hashCode() {
+		final int prime = 31;
+		int result = 1;
+		result = prime * result + id;
+		return result;
+	}
+	@Override
+	public boolean equals(Object obj) {
+		if (this == obj) {
+			return true;
+		}
+		if (obj == null) {
+			return false;
+		}
+		if (!(obj instanceof Note)) {
+			return false;
+		}
+		Note other = (Note) obj;
+		if (id != other.id) {
+			return false;
+		}
+		return true;
+	}
+	
+	
 	
 }
