+++ src/main/resources/templates/shoppingList/shoppingList.html
